requirements.txt
streamlit
pytesseract
pillow
pandas
asyncio
nltk
spacy
opencv-python
numpy
tqdm
fpdf
openai
python-docx
python-multipart
tiktoken
openpyxl
beautifulsoup4
requests
nest_asyncio
tabulate
